Use fixed sized ints where possible.
Based on clang-tidy google-runtime-int warnings.

+++ include/osmium/io/bzip2_compression.hpp
@@ -303,7 +303,7 @@ namespace osmium {
                         throw bzip2_error{message, result};
                     }
 
-                    output.resize(static_cast<unsigned long>(m_bzstream.next_out - output.data()));
+                    output.resize(static_cast<std::size_t>(m_bzstream.next_out - output.data()));
                 }
 
                 return output;
